% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceWeightsCalc.R
\name{abundanceWeightsCalc}
\alias{abundanceWeightsCalc}
\title{A function to calculate the weights based on the abundance
of points only.}
\usage{
abundanceWeightsCalc(
  pppx,
  pppy = NULL,
  square = FALSE,
  Acorr = FALSE,
  r = NULL,
  correction = NULL
)
}
\arguments{
\item{pppx}{A point pattern.}

\item{pppy}{A second point pattern for bivariate analyses.}

\item{square}{Scale weights to the number of pairs.}

\item{Acorr}{Correct for pattern area.}

\item{r}{A vector of distances to evaluate K at.}

\item{correction}{The edge correction to use. Refer to
\code{\link[spatstat.core]{Kest}} for options.
Only \code{border} is currently implemented, although NULL will also work
when no edge correction is required (e.g. in plus-sampling designs).}
}
\value{
A vector of weights of length r.
}
\description{
A function to calculate the weights based on the abundance
of points only.
}
\seealso{
Other RSPP weight calculations: 
\code{\link{kfuncWeightsCalc}()},
\code{\link{ratioWeightsCalc}()}
}
\concept{RSPP weight calculations}
