% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.klmer.R
\name{anova.klmer}
\alias{anova.klmer}
\title{Compare Models with and without a Parameter of Interest.}
\usage{
\method{anova}{klmer}(
  object,
  term,
  dists,
  nboot,
  maxit = 50,
  ncore = 1,
  cltype = "PSOCK",
  iseed = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{klmer}}}

\item{term}{The parameter of interest.}

\item{dists}{Distances at which to test the effect of term. Can be a list with
several distance ranges specified.}

\item{nboot}{Number of bootstrap interations to compute the null distribution.}

\item{maxit}{Maximum number of samples to try before giving up on refitting
a model at a given distance.}

\item{ncore}{Number of cpus to use.}

\item{cltype}{The cluster type, see \code{\link[parallel]{makeCluster}}}

\item{iseed}{The random number seed to use for repeatability.}

\item{...}{Additional arguments, currently not implemented.}
}
\value{
A list of class klmerAnova.
}
\description{
Compare Models with and without a Parameter of Interest.
}
