% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.klm.R
\name{autoplot.klm}
\alias{autoplot.klm}
\title{Builds a ggplot object to plot parameter estimates against distance}
\usage{
autoplot.klm(x, ...)
}
\arguments{
\item{x}{A \code{\link{klm}} object.}

\item{...}{additional arguments (currently unused).}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Builds a ggplot object to plot parameter estimates against distance
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmerHyper}()},
\code{\link{autoplot.klmerci}()},
\code{\link{makePlotData_klmerHyper}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{makePlotData_klm}()},
\code{\link{plot.klmerHyper}()},
\code{\link{plot.klmerci}()},
\code{\link{plot.klm}()}
}
\concept{RSPP plot functions}
