% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.klmerci.R
\name{autoplot.klmerci}
\alias{autoplot.klmerci}
\title{Builds a ggplot object to plot parameter estimates with their 
confidence intervals against distance}
\usage{
autoplot.klmerci(x, ...)
}
\arguments{
\item{x}{A klmerci object from \code{\link{confint.klmerHyper}}.}

\item{...}{additional arguments (currently unused).}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Builds a ggplot object to plot parameter estimates with their 
confidence intervals against distance
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmerHyper}()},
\code{\link{makePlotData_klmerHyper}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{plot.klmerHyper}()},
\code{\link{plot.klmerci}()}
}
\concept{RSPP plot functions}
