% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.klmerHyper.R
\name{autoplot.klmerHyper}
\alias{autoplot.klmerHyper}
\title{Builds a ggplot object to plot parameter estimates against distance}
\usage{
autoplot.klmerHyper(x, ...)
}
\arguments{
\item{x}{A \code{\link{klmerHyper}} object.}

\item{...}{additional arguments (currently unused).}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Builds a ggplot object to plot parameter estimates against distance
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmci}()},
\code{\link{autoplot.klmerci}()},
\code{\link{autoplot.klm}()},
\code{\link{makePlotData_klmci}()},
\code{\link{makePlotData_klmerHyper}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{makePlotData_klm}()},
\code{\link{plot.klmci}()},
\code{\link{plot.klmerHyper}()},
\code{\link{plot.klmerci}()},
\code{\link{plot.klm}()}
}
\concept{RSPP plot functions}
