% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.klmerHyper.R
\name{bootParallel.klmerHyper}
\alias{bootParallel.klmerHyper}
\title{Bootstrap on klmer models to obtain confidence intervals.}
\usage{
bootParallel.klmerHyper(mods, lin_comb, nboot, ncore = 1, cltype = "PSOCK",
  iseed = NULL)
}
\arguments{
\item{mods}{A \code{\link{klmerHyper}} object.}

\item{lin_comb}{linear combintaion of the fixed parameters}

\item{nboot}{number of bootstrap simulations.}

\item{ncore}{Number of cpus to use.}

\item{cltype}{Type of cluster to use.}

\item{iseed}{Random number seed.}
}
\value{
A bootstrap sample of predictions, their standard errors,
fixed effects and variance covariance matrix for all distances.
}
\description{
Bootstrap on klmer models to obtain confidence intervals.
}
