% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.klm.R
\name{bootstrap.klm}
\alias{bootstrap.klm}
\title{bootstrapping function for klm models to obtain confidence intervals.}
\usage{
bootstrap.klm(mods, lin_comb)
}
\arguments{
\item{mods}{A \code{\link{klm}} object.}

\item{lin_comb}{Linear combination of the fixed parameters}
}
\value{
A bootstrap sample of parameters, predictions, their standard errors,
fixed effects and variance covariance matrix for all distances.
}
\description{
bootstrapping function for klm models to obtain confidence intervals.
}
