% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.klmerHyper.R
\name{bootstrap.klmerHyper}
\alias{bootstrap.klmerHyper}
\title{Bootstrap on klmer models to obtain confidence intervals.}
\usage{
bootstrap.klmerHyper(mods, lin_comb, nboot, maxit = 10, ncore = 1,
  cltype = "PSOCK", iseed = NULL)
}
\arguments{
\item{mods}{A \code{\link{klmerHyper}} object.}

\item{lin_comb}{Linear combintaion of the fixed parameters}

\item{nboot}{Number of bootstrap simulations.}

\item{maxit}{Number of attempts to try fitting a bootstrap iteration before
giving up. Defaults to 10, which should be enough for most applications.}

\item{ncore}{Number of cpus to use.}

\item{cltype}{Type of cluster to use.}

\item{iseed}{Random number seed.}
}
\value{
A bootstrap sample of predictions, their standard errors,
fixed effects and variance covariance matrix for all distances.
}
\description{
Bootstrap on klmer models to obtain confidence intervals.
}
