% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.klm.R
\name{confint.klm}
\alias{confint.klm}
\title{Calculate confidence intervals on a klm object.}
\usage{
\method{confint}{klm}(
  object,
  parm,
  level = 0.95,
  newdata = NULL,
  lin_comb = NULL,
  nboot = 1,
  iseed = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{\link{klm}} object.}

\item{parm}{Parameter to get confidence interval for. Currently ignored.}

\item{level}{Desired confidence intervals.}

\item{newdata}{New data to make predictions on. Defaults to NULL whereupon 
fitted values are returned.}

\item{lin_comb}{Linear combination of the fixed parameters. Defaults
to NULL to extract it from the fitted model.}

\item{nboot}{Number of bootstrap simulations.}

\item{iseed}{Random number seed.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Returns the confidence intervals on an klm object.
}
\description{
Calculate confidence intervals on a klm object.
}
