% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.klmer.R
\name{confint.klmer}
\alias{confint.klmer}
\title{Calculate Bootstrap Confidence Intervals on a klmer object.}
\usage{
\method{confint}{klmer}(
  object,
  parm = NULL,
  level = 0.95,
  nboot = NULL,
  newdata = NULL,
  lin_comb = NULL,
  bootobj = NULL,
  ncore = 1,
  cltype = "PSOCK",
  iseed = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{\link{klmer}} object.}

\item{parm}{Parameter to get confidence interval for. Currently ignored.}

\item{level}{Desired confidence intervals.}

\item{nboot}{Number of bootstrap simulations.}

\item{newdata}{New data to make predictions on. Defaults to NULL whereupon
fitted values are returned.}

\item{lin_comb}{Linear combination of the fixed parameters}

\item{bootobj}{Bootstrap object run outside function. Defaults to NULL so
the bootstrap is run internally.}

\item{ncore}{Number of cpus to use.}

\item{cltype}{Type of cluster to use.}

\item{iseed}{Random number seed.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Returns the confidence intervals on an klmer object.
}
\description{
The function uses a semi-parameteric bootstrapping approach 
to compute confidence intervals on parameter estimates and predictions from 
based on a klmer model object. 

Currently, random effects are ignored in the computations of predcitions -
i.e. predictions are made for an "average" group but assuming that any
group effects are known (i.e., no uncertainty is added from not knowing
which group the prediction is for).
}
