% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.klmerHyper.R
\name{confint.klmerHyper}
\alias{confint.klmerHyper}
\title{Calculate Bootstrap Confidence Intervals on a klmerHyper object.}
\usage{
\method{confint}{klmerHyper}(mods, lin_comb = NULL, bootobj = NULL,
  confint = 0.95, nboot = NULL, ncore = 1, cltype = "PSOCK",
  iseed = NULL)
}
\arguments{
\item{mods}{A \code{\link{klmerHyper}} object.}

\item{lin_comb}{Linear combintaion of the fixed parameters}

\item{bootobj}{Bootstrap object run outside function. Defaults to NULL so
the bootstrap is run internally.}

\item{confint}{Desired confidence intervals.}

\item{nboot}{Number of bootstrap simulations.}

\item{ncore}{Number of cpus to use.}

\item{cltype}{Type of cluster to use.}

\item{iseed}{Random number seed.}
}
\value{
Returns the confidence intervals on an klmer object.
}
\description{
Calculate Bootstrap Confidence Intervals on a klmerHyper object.
}
\examples{

}
