% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateK.R
\name{generateK}
\alias{generateK}
\title{Generate Bootstrap Sample of K functions from models and resampled residuals.}
\usage{
generateK(mod, resids)
}
\arguments{
\item{mod}{A \code{\link{klmer}} model.}

\item{resids}{The sampled residuals and random effects.}
}
\value{
A K function.
}
\description{
Generate Bootstrap Sample of K functions from models and resampled residuals.
}
