% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPars.R
\name{getPars}
\alias{getPars}
\title{Extract Useful Parameters from Model and a Linear Contrast Matrix.}
\usage{
getPars(mod, lin_comb)
}
\arguments{
\item{mod}{Model to extract parameters from.}

\item{lin_comb}{Linear combination.}
}
\value{
A list with predictions, their standard errors, fixed effects and
the variance covariance matrix.
}
\description{
Extract Useful Parameters from Model and a Linear Contrast Matrix.
}
