% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KfuncWeightsCalc.R
\name{kfuncWeightsCalc}
\alias{kfuncWeightsCalc}
\title{Constructs the weights argument from a few simple rules.}
\usage{
kfuncWeightsCalc(
  pppx,
  pppy = NULL,
  r = NULL,
  correction = NULL,
  type = c("nx", "nx_A", "nx2", "nx2_A", "sqrtnxny", "nxny", "nxny_A", "sqrtnxny_A")
)
}
\arguments{
\item{pppx}{A point pattern.}

\item{pppy}{A second point pattern for bivariate analyses.}

\item{r}{A vector of distances to evaluate K at.}

\item{correction}{The edge correction to use - refer to
\code{\link{Kest}} for options.}

\item{type}{A string designating the types of weights to be calculated.}
}
\value{
A vector of weights for the K function with length equal to that of
r.
}
\description{
Constructs the weights argument from a few simple rules.
}
\details{
Uses the functions abundanceWeightsCalc and ratioWeightsCalc
to do the actual calculations.
}
\examples{
library(spatstat)
x <- rpoint(20)
y <- rpoint(20)
ratioWeightsCalc(x, y, r=seq(0, 0.25, 0.05), correction = "border")
}
\seealso{
Other RSPP weight calculations: 
\code{\link{abundanceWeightsCalc}()},
\code{\link{ratioWeightsCalc}()}
}
\concept{RSPP weight calculations}
