% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klm.R
\name{klm}
\alias{klm}
\title{Fits klm at specified distances from a hyperframe with kfunctions and covariates.}
\usage{
klm(
  formula,
  hyper,
  weights = NULL,
  weights_type = NULL,
  r,
  correction,
  ppx = NULL,
  minsamp = NA,
  remove_zero_weights = TRUE,
  na.action = "na.omit",
  printwarnings = TRUE
)
}
\arguments{
\item{formula}{Model formula following \code{\link[stats]{lm}} syntax, with
left hand side a k-function object.}

\item{hyper}{\code{\link[spatstat.geom]{hyperframe}} with  k functions, weights,
grouping factors and covariates}

\item{weights}{name (in quotes) of user defined weights column.}

\item{weights_type}{The type of weights to be used if weights is not defined above. 
Must be a type listed in \code{\link{KfuncWeightsCalc}}. Ignored if weights is defined.}

\item{r}{Distances at which to fit model}

\item{correction}{Edge correction. See \code{\link[spatstat.core]{Kest}}.}

\item{ppx}{A set of \code{\link[spatstat.geom]{ppp}} objects to use in calculating weights}

\item{minsamp}{Minimum number of points to include point pattern in model.
Not currently used.}

\item{remove_zero_weights}{Whether to remove point patterns with no focal points after 
edge corrections. If TRUE (the defaults) removes such rows with a warning.}

\item{na.action}{How to deal with missing data.}

\item{printwarnings}{Print warnings about distances with no variance?}
}
\value{
Model output of class klm
}
\description{
Fits klm at specified distances from a hyperframe with kfunctions and covariates.
}
