% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klm_i.R
\name{klm_i}
\alias{klm_i}
\title{## Function to do a linear regression of K functions}
\usage{
klm_i(formula, k, data, weights, na.action)
}
\arguments{
\item{formula}{A one-sided formula describing the model to be fit, following}

\item{k}{A list of Kfunctions.}

\item{data}{The covariates.}

\item{weights}{The weights vector, probably generated by KfuncWeightsCalc.
the lme4 syntax. See \code{\link[lme4]{lmer}}}

\item{na.action}{How to deal with missing values.
See \code{\link[stats]{na.action}}}
}
\value{
An object of class kfunclmer.
}
\description{
## Function to do a linear regression of K functions
}
