% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klmer_i.R
\name{klmer_i}
\alias{klmer_i}
\title{function to fit mixed effects model to k functions at a set distance}
\usage{
klmer_i(formula, k, data, weights, na.action = "na.omit")
}
\arguments{
\item{formula}{A one-sided formula describing the model to be fit, following}

\item{k}{A list of Kfunctions.}

\item{data}{The covariates.}

\item{weights}{The weights vector, probably generated by KfuncWeightsCalc.
the lme4 syntax. See \code{\link[lme4]{lmer}}}

\item{na.action}{How to deal with missing values.
See \code{\link[stats]{na.action}}}
}
\value{
An object of class kfunclmer.
}
\description{
function to fit mixed effects model to k functions at a set distance
}
