% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klmerHyper.R
\name{klmerHyper}
\alias{klmerHyper}
\title{Fits klmer at specified distances from a hyperframe with kfunctions and covariates.}
\usage{
klmerHyper(formula, hyper, r, correction, minsamp = NA,
  na.action = "na.omit", printwarnings = FALSE)
}
\arguments{
\item{formula}{Model formula following \code{\link[lme4]{lmer}} syntax}

\item{hyper}{\code{\link[spatstat]{hyperframe}} with  k functions, weights,
grouping factors and covariates}

\item{r}{Distances at which to fit model}

\item{correction}{Edge correction. See \code{\link[spatstat]{Kest}}.}

\item{minsamp}{Minimum number of points to include point pattern in model.
Not currently used.}

\item{na.action}{How to deal with missing data.}

\item{printwarnings}{Print warnings about distances with no variance?}
}
\value{
Model output of class klmerHyper
}
\description{
Fits klmer at specified distances from a hyperframe with kfunctions and covariates.
}
\examples{

}
