% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlotData_klm.R
\name{makePlotData_klm}
\alias{makePlotData_klm}
\title{Format fixed effect parameters from a klm object for easy plotting}
\usage{
makePlotData_klm(x)
}
\arguments{
\item{x}{A \code{\link{klm}} object.}
}
\value{
a \code{\link[tibble]{tibble}} of coefficients against distance.
}
\description{
Format fixed effect parameters from a klm object for easy plotting
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klm}()},
\code{\link{autoplot.klmci}()},
\code{\link{autoplot.klmer}()},
\code{\link{autoplot.klmerci}()},
\code{\link{makePlotData_klmci}()},
\code{\link{makePlotData_klmer}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{plot.klm}()},
\code{\link{plot.klmci}()},
\code{\link{plot.klmer}()},
\code{\link{plot.klmerci}()}
}
\concept{RSPP plot functions}
