% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlotData_klmci.R
\name{makePlotData_klmci}
\alias{makePlotData_klmci}
\title{Format a klmci object for plotting of parameters and confidence intervals
from a bootstrap}
\usage{
makePlotData_klmci(x)
}
\arguments{
\item{x}{A klmci object from \code{\link{confint.klm}}}
}
\value{
a \code{\link[tibble]{tibble}} of fixed effects against distance.
}
\description{
Format a klmci object for plotting of parameters and confidence intervals
from a bootstrap
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmci}()},
\code{\link{autoplot.klmerHyper}()},
\code{\link{autoplot.klmerci}()},
\code{\link{autoplot.klm}()},
\code{\link{makePlotData_klmerHyper}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{makePlotData_klm}()},
\code{\link{plot.klmci}()},
\code{\link{plot.klmerHyper}()},
\code{\link{plot.klmerci}()},
\code{\link{plot.klm}()}
}
\concept{RSPP plot functions}
