% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modCompare.R
\name{modCompare}
\alias{modCompare}
\title{Compares the Deviances of Two Models.}
\usage{
modCompare(modsH1, modsH0)
}
\arguments{
\item{modsH1}{A more complex model of class \code{\link{klmer}}}

\item{modsH0}{A simpler (null) model of class \code{\link{klmer}}}
}
\value{
The difference in deviances between the null and complex model.
}
\description{
Compares the Deviances of Two Models.
}
