% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modCompareBoot.R
\name{modCompareBoot}
\alias{modCompareBoot}
\title{Simulate Deviance Differences between klmer Models Under a Null Hypothesis.}
\usage{
modCompareBoot(modsH1, modsH0, resids, maxit)
}
\arguments{
\item{modsH1}{The full model of class \code{\link{klmerHyper}}}

\item{modsH0}{The null model of class \code{\link{klmerHyper}}}

\item{resids}{Exchangeable residuals from modsH1}

\item{maxit}{The maximum number of iterations to try before giving up.}
}
\value{
The vector of simulated deviances for all distances.
}
\description{
Simulate Deviance Differences between klmer Models Under a Null Hypothesis.
}
