% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.klm.R
\name{plot.klm}
\alias{plot.klm}
\title{Plot a klm object. Plots parameter estimates against distance}
\usage{
\method{plot}{klm}(x, ...)
}
\arguments{
\item{x}{A \code{\link{klm}} object.}

\item{...}{additional arguments (currently unused).}
}
\value{
a ggplot object
}
\description{
Plot a klm object. Plots parameter estimates against distance
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klm}()},
\code{\link{autoplot.klmci}()},
\code{\link{autoplot.klmer}()},
\code{\link{autoplot.klmerci}()},
\code{\link{makePlotData_klm}()},
\code{\link{makePlotData_klmci}()},
\code{\link{makePlotData_klmer}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{plot.klmci}()},
\code{\link{plot.klmer}()},
\code{\link{plot.klmerci}()}
}
\concept{RSPP plot functions}
