% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.klmci.R
\name{plot.klmci}
\alias{plot.klmci}
\title{Plot a klmci object  with parameters and confidence intervals
from a bootstrap}
\usage{
\method{plot}{klmci}(x, ...)
}
\arguments{
\item{x}{A klmci object from \code{\link{confint.klm}}.}

\item{...}{additional arguments (currently unused).}
}
\value{
a ggplot object
}
\description{
Plot a klmci object  with parameters and confidence intervals
from a bootstrap
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmci}()},
\code{\link{autoplot.klmerci}()},
\code{\link{autoplot.klmer}()},
\code{\link{autoplot.klm}()},
\code{\link{makePlotData_klmci}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{makePlotData_klmer}()},
\code{\link{makePlotData_klm}()},
\code{\link{plot.klmerci}()},
\code{\link{plot.klmer}()},
\code{\link{plot.klm}()}
}
\concept{RSPP plot functions}
