% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.klmer.R
\name{plot.klmer}
\alias{plot.klmer}
\title{Plot a klmer object. Plots parameter
estimates distance}
\usage{
\method{plot}{klmer}(x, ...)
}
\arguments{
\item{x}{A \code{\link{klmer}} object.}

\item{...}{additional arguments (currently unused).}
}
\value{
a ggplot object
}
\description{
Plot a klmer object. Plots parameter
estimates distance
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmci}()},
\code{\link{autoplot.klmerci}()},
\code{\link{autoplot.klmer}()},
\code{\link{autoplot.klm}()},
\code{\link{makePlotData_klmci}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{makePlotData_klmer}()},
\code{\link{makePlotData_klm}()},
\code{\link{plot.klmci}()},
\code{\link{plot.klmerci}()},
\code{\link{plot.klm}()}
}
\concept{RSPP plot functions}
