% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.klmerci.R
\name{plot.klmerci}
\alias{plot.klmerci}
\title{Plot a klmerci object  with parameters and confidence intervals
from a bootstrap}
\usage{
\method{plot}{klmerci}(x, ...)
}
\arguments{
\item{x}{A klmerci object from \code{\link{confint.klmerHyper}}.}

\item{...}{additional arguments (currently unused).}
}
\value{
a ggplot object
}
\description{
Plot a klmerci object  with parameters and confidence intervals
from a bootstrap
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmerHyper}()},
\code{\link{autoplot.klmerci}()},
\code{\link{makePlotData_klmerHyper}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{plot.klmerHyper}()}
}
\concept{RSPP plot functions}
