% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.klmerHyper.R
\name{plot.klmerHyper}
\alias{plot.klmerHyper}
\title{Plot a klmerHyper object. Plots parameter
estimates distance}
\usage{
\method{plot}{klmerHyper}(x, ...)
}
\arguments{
\item{x}{A klmerHyper object}

\item{...}{additional arguments (currently unused).}
}
\value{
a ggplot object
}
\description{
Plot a klmerHyper object. Plots parameter
estimates distance
}
\concept{RSPP plot functions}
