% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.klm.R
\name{print.klm}
\alias{print.klm}
\title{Print klm Objects}
\usage{
\method{print}{klm}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{...}{Additional arguments to print. Not implemented yet.}
}
\description{
Print klm Objects
}
