% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.klmci.R
\name{print.klmci}
\alias{print.klmci}
\title{Print klm confidence interval objects}
\usage{
\method{print}{klmci}(x, type = "pars", ...)
}
\arguments{
\item{x}{A klmci object.}

\item{type}{Print parameters, (pars, the default), or predictions (predictions)}

\item{...}{Additional arguments to print. Not implemented yet.}
}
\value{
No return value.
}
\description{
Print klm confidence interval objects
}
