% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.klmerAnova.R
\name{print.klmerAnova}
\alias{print.klmerAnova}
\title{Print a klmerAnova Object}
\usage{
\method{print}{klmerAnova}(obj, stat = T)
}
\arguments{
\item{obj}{An object of class \code{\link{anova.klmerHyper}}}
}
\value{
Prints to screen.
}
\description{
Print a klmerAnova Object
}
