% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.klmerci.R
\name{print.klmerci}
\alias{print.klmerci}
\title{Print klmer confidence interval objects}
\usage{
\method{print}{klmerci}(obj, type = "pars")
}
\arguments{
\item{obj}{A klmerHyper object.}

\item{type}{Print parameters, (pars, the default), or predictions (preds)}
}
\value{
No return value.
}
\description{
Print klmer confidence interval objects
}
