% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranefRandomise.R
\name{ranefRandomise}
\alias{ranefRandomise}
\title{Randomise BLUPs After Correcting for Shrinkage.}
\usage{
ranefRandomise(mods, ranefs)
}
\arguments{
\item{mods}{A list of models returned by \code{\link{klmer}}}

\item{ranefs}{A list of random effects corresponding to mods.}
}
\value{
A list of resampled random effects.
}
\description{
Randomise BLUPs After Correcting for Shrinkage.
}
