% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioWeightsCalc.R
\name{ratioWeightsCalc}
\alias{ratioWeightsCalc}
\title{Takes a Point Pattern and Calculates the Denominator of the K Function.}
\usage{
ratioWeightsCalc(pppx, pppy = NULL, r = NULL, correction = "border")
}
\arguments{
\item{pppx}{A point pattern.}

\item{pppy}{A second point pattern for bivariate analyses.}

\item{r}{A vector of distances to evaluate K at.}

\item{correction}{The edge correction to use - refer to Kest for options.}
}
\value{
A vector with the denominator of the K functions for each point
}
\description{
Takes a Point Pattern and Calculates the Denominator of the K Function.
}
\examples{
x <- spatstat::rpoint(20)
y <- spatstat::rpoint(20)
ratioWeightsCalc(x, y, r=seq(0, 0.2, 0.05), correction = "border")

}
\seealso{
Other RSPP weight calculations: \code{\link{abundanceWeightsCalc}},
  \code{\link{kfuncWeightsCalc}}
}
