% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.klmer.R
\name{refit.klmer}
\alias{refit.klmer}
\title{Refits an klmer model to the Bootstrapped K Function Data}
\usage{
\method{refit}{klmer}(object, newresp, ...)
}
\arguments{
\item{object}{Original \code{\link{klmer}} model fitted to observed data.}

\item{newresp}{Simulated K functions.}

\item{...}{Other arguments passed to methods. Currently ignored.}
}
\value{
A klmer model fitted to the randomised data.
}
\description{
Refits an klmer model to the Bootstrapped K Function Data
}
