% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.klmerHyper.R
\name{refit.klmerHyper}
\alias{refit.klmerHyper}
\title{Refits an klmer Model to the Bootstrapped K Function Data}
\usage{
\method{refit}{klmerHyper}(mods, newK)
}
\arguments{
\item{mod}{Original klmer model fitted to observed data.}

\item{res_r}{randomised residuals.}
}
\value{
An klmer model fitted to the randomised data.
}
\description{
Refits an klmer Model to the Bootstrapped K Function Data
}
\note{
This is probably the time consuming step. Might be worth setting the
parameters of mod as starting values and increasing tolerances several fold.
We are only interested in the paramter estimates here, not their
uncertainties. Speeding up the code will make it feasible to increase
bootstraps, which is where better estimates of uncertainty will come from.
}
