% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitML.klmer.R
\name{refitMLklmer}
\alias{refitMLklmer}
\title{Function to Convert klmer Objects from REML to ML}
\usage{
refitMLklmer(mods)
}
\arguments{
\item{mods}{model to be converted}
}
\value{
A klmer object with models fitted via ML
}
\description{
Function to Convert klmer Objects from REML to ML
}
