% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitML.klmerHyper.R
\name{refitMLklmerHyper}
\alias{refitMLklmerHyper}
\title{Function to Convert klmer Objects from REML to ML}
\usage{
refitMLklmerHyper(mods)
}
\arguments{
\item{mods}{model to be converted}
}
\value{
A klmerHyper object with models fitted via ML
}
\description{
Function to Convert klmer Objects from REML to ML
}
