% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleblups.R
\name{rescaleblups}
\alias{rescaleblups}
\title{Rescale BLUPs to be exchangable.}
\usage{
rescaleblups(U, Sigma, forcePD = TRUE)
}
\arguments{
\item{U}{Random effects.}

\item{Sigma}{The estimated variance-covariance matrix.}

\item{forcePD}{Force all matrices to be positive definite.}
}
\value{
A list of corrected BLUPs
}
\description{
Rescale BLUPs to be exchangable.
}
