% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residHomogenise.klm.R
\name{residHomogenise.klm}
\alias{residHomogenise.klm}
\title{Homogenises residuals from a lm object to make them exchangable.}
\usage{
residHomogenise.klm(mod)
}
\arguments{
\item{mod}{lm model}
}
\value{
List of residuals and random effects that are exchangable.
}
\description{
Homogenises residuals from a lm object to make them exchangable.
}
