% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residHomogenise.R
\name{residHomogenise}
\alias{residHomogenise}
\title{Homogenises residuals from a lmer object to make them exchangable.}
\usage{
residHomogenise(mods)
}
\arguments{
\item{mod}{An lmer model}
}
\value{
List of residuals and random effects that are exchangable.
}
\description{
Homogenises residuals from a lmer object to make them exchangable.
}
