% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residRandomise.R
\name{residRandomise}
\alias{residRandomise}
\title{Randomise Homogenised Level 1 Residuals}
\usage{
residRandomise(resids_e, mods)
}
\arguments{
\item{resids_e}{= Homogenised level 1 residuals from
\code{\link{residualHomogenise.klmer}}.}

\item{mods}{klmer models.}
}
\value{
A list of randomised level 1 residuals.
}
\description{
Randomise Homogenised Level 1 Residuals
}
