% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualHomogenise.klmer.R
\name{residualHomogenise.klmer}
\alias{residualHomogenise.klmer}
\title{Homogenises residuals from a lmer object to make them exchangable.}
\usage{
residualHomogenise.klmer(mod)
}
\arguments{
\item{mod}{An lmer model}
}
\value{
List of residuals and random effects that are exchangable.
}
\description{
Homogenises residuals from a lmer object to make them exchangable.
}
