% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.klmer.R
\name{simulate.klmer}
\alias{simulate.klmer}
\title{Semi-Parametric Bootstrap sample form a klmer object.}
\usage{
\method{simulate}{klmer}(object, nsim = 1, seed = NULL, resids, ...)
}
\arguments{
\item{object}{Models to be simulated from.}

\item{nsim}{Number of simulations (ignored)}

\item{seed}{Seed for simulations (currently ignored)}

\item{resids}{The residuals to be resampled.}

\item{...}{Additional parameters for simulate method.}
}
\value{
A list of simulated Kfunctions for all distances in klmer object
object.
}
\description{
Semi-Parametric Bootstrap sample form a klmer object.
}
