% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.klmerHyper.R
\name{simulate.klmerHyper}
\alias{simulate.klmerHyper}
\title{Semi-Parametric Bootstrap sample form a klmerHyper object.}
\usage{
\method{simulate}{klmerHyper}(mods, resids)
}
\arguments{
\item{mods}{list of klmer models, usually from \code{\link{klmerHyper}}}
}
\value{
A list of simulated Kfunctions for all distances in klmerHyper
object.
}
\description{
Semi-Parametric Bootstrap sample form a klmerHyper object.
}
