% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.klmer.R
\name{update.klmer}
\alias{update.klmer}
\title{Updates klmer Objects by Dropping Specified Term from Model}
\usage{
\method{update}{klmer}(object, term, ...)
}
\arguments{
\item{object}{klmer object.}

\item{term}{Term to be tested.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Returns a klmer object with the term of interest dropped.
}
\description{
Updates klmer Objects by Dropping Specified Term from Model
}
