% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.klmerHyper.R
\name{update.klmerHyper}
\alias{update.klmerHyper}
\title{Updates klmerHyper Objects by Dropping Specified Term from Model}
\usage{
\method{update}{klmerHyper}(mods, term)
}
\arguments{
\item{mods}{klmerHyper object.}

\item{term}{Term to be tested.}
}
\value{
Returns a klmerHyper object with the term of interest dropped.
}
\description{
Updates klmerHyper Objects by Dropping Specified Term from Model
}
