% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.klmci.R
\name{autoplot.klmci}
\alias{autoplot.klmci}
\title{Builds a ggplot object to plot parameter estimates with their 
confidence intervals against distance}
\usage{
autoplot.klmci(x, ...)
}
\arguments{
\item{x}{A klmci object from \code{\link{confint.klm}}.}

\item{...}{additional arguments (currently unused).}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Builds a ggplot object to plot parameter estimates with their 
confidence intervals against distance
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klm}()},
\code{\link{autoplot.klmer}()},
\code{\link{autoplot.klmerci}()},
\code{\link{makePlotData_klm}()},
\code{\link{makePlotData_klmci}()},
\code{\link{makePlotData_klmer}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{plot.klm}()},
\code{\link{plot.klmci}()},
\code{\link{plot.klmer}()},
\code{\link{plot.klmerci}()}
}
\concept{RSPP plot functions}
