% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.klm.R
\name{confint.klm}
\alias{confint.klm}
\title{Calculate confidence intervals on a klm object.}
\usage{
\method{confint}{klm}(object, lincomb, nsim = 1, level, iseed = NULL)
}
\arguments{
\item{object}{Bootstrap object run outside function. Defaults to NULL so
the bootstrap is run internally.}

\item{level}{Desired confidence intervals.}

\item{iseed}{Random number seed.}

\item{lin_comb}{Linear combintaion of the fixed parameters}

\item{nboot}{Number of bootstrap simulations.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Returns the confidence intervals on an klm object.
}
\description{
Calculate confidence intervals on a klm object.
}
