% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.klmerHyper.R
\name{confint.klmerHyper}
\alias{confint.klmerHyper}
\title{Calculate Bootstrap Confidence Intervals on a klmerHyper object.}
\usage{
\method{confint}{klmerHyper}(
  object,
  parm = NULL,
  level = 0.95,
  lin_comb = NULL,
  bootobj = NULL,
  nboot = NULL,
  ncore = 1,
  cltype = "PSOCK",
  iseed = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{\link{klmerHyper}} object.}

\item{parm}{Parameter to get confidence interval for. Currently ignored.}

\item{level}{Desired confidence intervals.}

\item{lin_comb}{Linear combintaion of the fixed parameters}

\item{bootobj}{Bootstrap object run outside function. Defaults to NULL so
the bootstrap is run internally.}

\item{nboot}{Number of bootstrap simulations.}

\item{ncore}{Number of cpus to use.}

\item{cltype}{Type of cluster to use.}

\item{iseed}{Random number seed.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Returns the confidence intervals on an klmerHyper object.
}
\description{
Calculate Bootstrap Confidence Intervals on a klmerHyper object.
}
