% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klmer.R
\name{klmer}
\alias{klmer}
\title{function to fit mixed effects model to k functions at a set distance}
\usage{
klmer(formula, k, data, weights, na.action = "na.omit")
}
\arguments{
\item{formula}{A one-sided formula describing the model to be fit, following}

\item{k}{A list of Kfunctions.}

\item{data}{The covariates.}

\item{weights}{The weights vector, probably generated by KfuncWeightsCalc.
the lme4 syntax. See \code{\link[lme4]{lmer}}}

\item{na.action}{How to deal with missing values.
See \code{\link[stats]{na.action}}}
}
\value{
An object of class kfunclmer.
}
\description{
function to fit mixed effects model to k functions at a set distance
}
\examples{

library(spatstat)
pppx <- replicate(50, rpoispp(runif(1, 10, 100)), simplify=FALSE)
r <- seq(0, 0.25, 0.05)
k <- lapply(pppx, Kest, r=r, correction='border')
k <- sapply(k, function(x) x[['border']][2])
weights <- lapply(pppx, kfuncWeightsCalc, r=r,
correction="border", type="nx_A")
weights <- sapply(weights, function(x) x[[2]])
dat <- data.frame(x=runif(50, 0, 10), gr = sample(1:10, 50, replace=TRUE))
formula = ~ x + (1|gr)
klmer(~x + (1|gr), k=k, data=dat, weights=weights)
}
