% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlotData_klm.R
\name{makePlotData_klm}
\alias{makePlotData_klm}
\title{Format fixed effect parameters from a klmerHyper object for easy plotting}
\usage{
makePlotData_klm(x)
}
\arguments{
\item{x}{A \code{\link{klm}} object.}
}
\value{
a \code{\link[tibble]{tibble}} of fixed effects against distance.
}
\description{
Format fixed effect parameters from a klmerHyper object for easy plotting
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmerHyper}()},
\code{\link{autoplot.klmerci}()},
\code{\link{autoplot.klm}()},
\code{\link{makePlotData_klmerHyper}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{plot.klmerHyper}()},
\code{\link{plot.klmerci}()},
\code{\link{plot.klm}()}
}
\concept{RSPP plot functions}
