% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlotData_klmerci.R
\name{makePlotData_klmerci}
\alias{makePlotData_klmerci}
\title{Format a klmerci object for plotting of parameters and confidence intervals
from a bootstrap}
\usage{
makePlotData_klmerci(x)
}
\arguments{
\item{x}{A klmerci object from \code{\link{confint.klmerHyper}}}
}
\value{
a \code{\link[tibble]{tibble}} of fixed effects against distance.
}
\description{
Format a klmerci object for plotting of parameters and confidence intervals
from a bootstrap
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmci}()},
\code{\link{autoplot.klmerHyper}()},
\code{\link{autoplot.klmerci}()},
\code{\link{autoplot.klm}()},
\code{\link{makePlotData_klmci}()},
\code{\link{makePlotData_klmerHyper}()},
\code{\link{makePlotData_klm}()},
\code{\link{plot.klmci}()},
\code{\link{plot.klmerHyper}()},
\code{\link{plot.klmerci}()},
\code{\link{plot.klm}()}
}
\concept{RSPP plot functions}
