% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.klm.R
\name{plot.klm}
\alias{plot.klm}
\title{Plot a klmerHyper object. Plots parameter
estimates distance}
\usage{
\method{plot}{klm}(x, ...)
}
\arguments{
\item{x}{A \code{\link{klm}} object.}

\item{...}{additional arguments (currently unused).}
}
\value{
a ggplot object
}
\description{
Plot a klmerHyper object. Plots parameter
estimates distance
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmerHyper}()},
\code{\link{autoplot.klmerci}()},
\code{\link{autoplot.klm}()},
\code{\link{makePlotData_klmerHyper}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{makePlotData_klm}()},
\code{\link{plot.klmerHyper}()},
\code{\link{plot.klmerci}()}
}
\concept{RSPP plot functions}
