% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.klmerHyper.R
\name{plot.klmerHyper}
\alias{plot.klmerHyper}
\title{Plot a klmerHyper object. Plots parameter
estimates distance}
\usage{
\method{plot}{klmerHyper}(x, ...)
}
\arguments{
\item{x}{A \code{\link{klmerHyper}} object.}

\item{...}{additional arguments (currently unused).}
}
\value{
a ggplot object
}
\description{
Plot a klmerHyper object. Plots parameter
estimates distance
}
\seealso{
Other RSPP plot functions: 
\code{\link{autoplot.klmerHyper}()},
\code{\link{autoplot.klmerci}()},
\code{\link{autoplot.klm}()},
\code{\link{makePlotData_klmerHyper}()},
\code{\link{makePlotData_klmerci}()},
\code{\link{makePlotData_klm}()},
\code{\link{plot.klmerci}()},
\code{\link{plot.klm}()}
}
\concept{RSPP plot functions}
