% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.klmerAnova.R
\name{print.klmerAnova}
\alias{print.klmerAnova}
\title{Print a klmerAnova Object}
\usage{
\method{print}{klmerAnova}(x, stat = T, ...)
}
\arguments{
\item{x}{An object of class \code{\link{anova.klmerHyper}}}

\item{stat}{Statistic to print. Options are T (default) or D.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Prints to screen.
}
\description{
Print a klmerAnova Object
}
