% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.klmerci.R
\name{print.klmerci}
\alias{print.klmerci}
\title{Print klmer confidence interval objects}
\usage{
\method{print}{klmerci}(x, type = "pars", ...)
}
\arguments{
\item{x}{A klmer object.}

\item{type}{Print parameters, (pars, the default), or predictions (predictions)}

\item{...}{Additional arguments to print. Not implemented yet.}
}
\value{
No return value.
}
\description{
Print klmer confidence interval objects
}
