% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.klmer.R
\name{refit.klmer}
\alias{refit.klmer}
\title{Refits an klmer model to the Bootstrapped K Function Data}
\usage{
\method{refit}{klmer}(mods, newK)
}
\arguments{
\item{mods}{Original \code{\link{klmer}} model fitted to observed data.}

\item{newK}{Simulated K functions.}
}
\value{
A klmer model fitted to the randomised data.
}
\description{
Refits an klmer model to the Bootstrapped K Function Data
}
