% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.klmerHyper.R
\name{refit.klmerHyper}
\alias{refit.klmerHyper}
\title{Refits an klmer Model to the Bootstrapped K Function Data}
\usage{
\method{refit}{klmerHyper}(mods, newK)
}
\arguments{
\item{mods}{Original \code{\link{klmerHyper}} model fitted to observed data.}

\item{newK}{Simulated K functions.}
}
\value{
A klmerHyper model fitted to the randomised data.
}
\description{
Refits an klmer Model to the Bootstrapped K Function Data
}
